IKHSS7-45228: Add Offset to Battery Health Calculation
The Meter Offset that is used to smooth the low end of the Battery Meter
does not get accounted for in the Battery Health Calculation which leads
to Aged Capacity that diminished by up to twice the Error in the System.
This change needs to feed the Offset back to Battery Health Daemon so it
can account for it.

Change-Id: I2a63a1d8f18bf2ad3175e2c0f34654116b2d5612
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Change-Id: I2846d3bd6532355597af502f12247422d640e7e1
Reviewed-on: http://gerrit.pcs.mot.com/456123
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>